Workflows: Add CI to export Linux build
This adds CI to publish Linux builds. How it works:

- Export ROTA from source using Godot Engine
- Upload that export as a build artifact
- On GitHub releases, attach that build artifact to the release

Eventually this could be expanded to build a Flatpak, a prerequisite to
getting it onto Flathub.

I intended to use the [godot-ci GitHub
Action](https://github.com/marketplace/actions/godot-ci) for CI, but it
doesn't currently support Godot 3.6 (see: abarichello/godot-ci#155). So
for now, I'm using
[docker.io/smks/godot-ci](https://hub.docker.com/r/smks/godot-ci) which
is a container image forked from the former and just updated for 3.6.